Refactored Window

+++ src/Core/Common/DurationToString.cpp
@@ -2,7 +2,7 @@
 
 namespace core::common {
 namespace {
-	int countDigits(size_t const x) {
+	int countDigits(u64 const x) {
 		if (x >= 10000000000) {
 			if (x >= 100000000000000) {
 				if (x >= 10000000000000000) {
@@ -52,7 +52,7 @@ namespace {
 		return 1;
 	}
 
-	void writeDigits(char*& buff, size_t number) noexcept {
+	void writeDigits(char*& buff, u64 number) noexcept {
 		int n = countDigits(number);
 		buff += n;
 		char* ptr = buff;
@@ -62,9 +62,9 @@ namespace {
 		}
 	}
 
-	template<size_t N>
-	void writeDigits(char*& buff, size_t number) noexcept {
-		size_t n = N;
+	template<usize N>
+	void writeDigits(char*& buff, u64 number) noexcept {
+		usize n = N;
 		while (n-- > 0) {
 			buff[n] = '0' + number % 10;
 			number /= 10;
@@ -77,20 +77,20 @@ namespace {
 		static thread_local char s_data[64];
 		char* ptr = s_data;
 		bool hasHigherUnit = false;
-		if (size_t const days = duration.days()) {
+		if (usize const days = duration.days()) {
 			constexpr static char const* const DAYS_STR[] = { " day, \0\0", " days, \0" };
 			writeDigits(ptr, days);
 			bool const moreThanOneDay = days > 1;
-			*reinterpret_cast<uint64_t*>(ptr) = *reinterpret_cast<uint64_t const*>(DAYS_STR[moreThanOneDay]);
+			*reinterpret_cast<u64*>(ptr) = *reinterpret_cast<u64 const*>(DAYS_STR[moreThanOneDay]);
 			ptr += 6 + moreThanOneDay;
 			hasHigherUnit = true;
 		}
-		if (size_t const hours = duration.hours(); hours || hasHigherUnit) {
+		if (usize const hours = duration.hours(); hours || hasHigherUnit) {
 			writeDigits<2>(ptr, hours);
 			*(ptr++) = ':';
 			hasHigherUnit = true;
 		}
-		if (size_t const minutes = duration.minutes(); minutes || hasHigherUnit) {
+		if (usize const minutes = duration.minutes(); minutes || hasHigherUnit) {
 			writeDigits<2>(ptr, minutes);
 			*(ptr++) = ':';
 			hasHigherUnit = true;
@@ -103,6 +103,6 @@ namespace {
 		*(ptr++) = '\'';
 		writeDigits<3>(ptr, duration.ns % 1'000);
 
-		return std::string_view(s_data, static_cast<size_t>(ptr - s_data));
+		return std::string_view(s_data, static_cast<usize>(ptr - s_data));
 	}
 } // namespace core::common
